+++ src/keybinds/config.rs
@@ -74,9 +74,9 @@ pub struct LogTabKeybindsConfig {
     pub copy_change_id: Option<Keybind>,
     pub copy_rev: Option<Keybind>,
     pub rebase: Option<Keybind>,
+    pub rebase_branch: Option<Keybind>,
 
     pub push: Option<Keybind>,
-    pub push_all: Option<Keybind>,
     pub fetch: Option<Keybind>,
     pub fetch_all: Option<Keybind>,